fix(user): moved User model to models directory and also name space changed in various file
