+++ app/Http/Controllers/TagController.php
@@ -11,7 +11,6 @@ use Exception;
 use Illuminate\Contracts\View\Factory;
 use Illuminate\Http\JsonResponse;
 use Illuminate\Http\Request;
-use Illuminate\Http\Response;
 use Illuminate\View\View;
 
 class TagController extends AppBaseController
