+++ app/Http/Requests/UpdateUserRequest.php
@@ -10,7 +10,7 @@
 
 namespace App\Http\Requests;
 
-use App\User;
+use App\Models\User;
 use Illuminate\Foundation\Http\FormRequest;
 class UpdateUserRequest extends FormRequest
 {
