+++ app/Models/TimeEntry.php
@@ -2,7 +2,6 @@
 
 namespace App\Models;
 
-use App\User;
 use Eloquent as Model;
 use Illuminate\Database\Eloquent\SoftDeletes;
 
@@ -15,19 +14,25 @@ use Illuminate\Database\Eloquent\SoftDeletes;
  * @property int $user_id
  * @property string|null $start_time
  * @property string|null $end_time
- * @property int|null $duration
- * @property string|null $note
+ * @property int $duration
+ * @property string $note
  * @property int|null $deleted_by
  * @property \Illuminate\Support\Carbon|null $created_at
  * @property \Illuminate\Support\Carbon|null $updated_at
- * @property-read \App\Models\Task $task
- * @property-read \App\User $user
+ * @property \Illuminate\Support\Carbon|null $deleted_at
  * @property-read \App\Models\ActivityType $activityType
+ * @property-read \App\Models\Task $task
+ * @property-read \App\Models\User $user
+ * @method static bool|null forceDelete()
  * @method static \Illuminate\Database\Eloquent\Builder|\App\Models\TimeEntry newModelQuery()
  * @method static \Illuminate\Database\Eloquent\Builder|\App\Models\TimeEntry newQuery()
+ * @method static \Illuminate\Database\Query\Builder|\App\Models\TimeEntry onlyTrashed()
  * @method static \Illuminate\Database\Eloquent\Builder|\App\Models\TimeEntry query()
+ * @method static bool|null restore()
  * @method static \Illuminate\Database\Eloquent\Builder|\App\Models\TimeEntry whereActivityTypeId($value)
  * @method static \Illuminate\Database\Eloquent\Builder|\App\Models\TimeEntry whereCreatedAt($value)
+ * @method static \Illuminate\Database\Eloquent\Builder|\App\Models\TimeEntry whereDeletedAt($value)
+ * @method static \Illuminate\Database\Eloquent\Builder|\App\Models\TimeEntry whereDeletedBy($value)
  * @method static \Illuminate\Database\Eloquent\Builder|\App\Models\TimeEntry whereDuration($value)
  * @method static \Illuminate\Database\Eloquent\Builder|\App\Models\TimeEntry whereEndTime($value)
  * @method static \Illuminate\Database\Eloquent\Builder|\App\Models\TimeEntry whereId($value)
@@ -36,15 +41,9 @@ use Illuminate\Database\Eloquent\SoftDeletes;
  * @method static \Illuminate\Database\Eloquent\Builder|\App\Models\TimeEntry whereTaskId($value)
  * @method static \Illuminate\Database\Eloquent\Builder|\App\Models\TimeEntry whereUpdatedAt($value)
  * @method static \Illuminate\Database\Eloquent\Builder|\App\Models\TimeEntry whereUserId($value)
- * @method static \Illuminate\Database\Eloquent\Builder|\App\Models\TimeEntry whereDeletedBy($value)
- * @mixin \Eloquent
- * @property \Illuminate\Support\Carbon|null $deleted_at
- * @method static bool|null forceDelete()
- * @method static \Illuminate\Database\Query\Builder|\App\Models\TimeEntry onlyTrashed()
- * @method static bool|null restore()
- * @method static \Illuminate\Database\Eloquent\Builder|\App\Models\TimeEntry whereDeletedAt($value)
  * @method static \Illuminate\Database\Query\Builder|\App\Models\TimeEntry withTrashed()
  * @method static \Illuminate\Database\Query\Builder|\App\Models\TimeEntry withoutTrashed()
+ * @mixin \Eloquent
  */
 class TimeEntry extends Model
 {
